further flatten the package tree
