+++ contrib/google.golang.org/grpc/appsec.go
@@ -10,10 +10,11 @@ import (
 
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace"
 	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/dyngo"
-	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/dyngo/instrumentation"
-	grpcsec "gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/dyngo/instrumentation/grpcsec/emitter"
-	httpsec "gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/dyngo/instrumentation/httpsec/emitter"
-	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/dyngo/instrumentation/sharedsec/emitter"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/emitter/grpcsec"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/emitter/sharedsec"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/trace"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/trace/grpctrace"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/trace/httptrace"
 	"gopkg.in/DataDog/dd-trace-go.v1/internal/log"
 
 	"github.com/DataDog/appsec-internal-go/netip"
@@ -26,14 +27,14 @@ import (
 
 // UnaryHandler wrapper to use when AppSec is enabled to monitor its execution.
 func appsecUnaryHandlerMiddleware(span ddtrace.Span, handler grpc.UnaryHandler) grpc.UnaryHandler {
-	instrumentation.SetAppSecEnabledTags(span)
+	trace.SetAppSecEnabledTags(span)
 	return func(ctx context.Context, req interface{}) (interface{}, error) {
 		var err error
 		var blocked bool
 		md, _ := metadata.FromIncomingContext(ctx)
 		clientIP := setClientIP(ctx, span, md)
 		args := grpcsec.HandlerOperationArgs{Metadata: md, ClientIP: clientIP}
-		ctx, op := grpcsec.StartHandlerOperation(ctx, args, nil, dyngo.NewDataListener(func(a *emitter.Action) {
+		ctx, op := grpcsec.StartHandlerOperation(ctx, args, nil, dyngo.NewDataListener(func(a *sharedsec.Action) {
 			code, e := a.GRPC()(md)
 			blocked = a.Blocking()
 			err = status.Error(codes.Code(code), e.Error())
@@ -41,12 +42,12 @@ func appsecUnaryHandlerMiddleware(span ddtrace.Span, handler grpc.UnaryHandler) 
 		defer func() {
 			events := op.Finish(grpcsec.HandlerOperationRes{})
 			if blocked {
-				op.AddTag(instrumentation.BlockedRequestTag, true)
+				op.AddTag(trace.BlockedRequestTag, true)
 			}
-			grpcsec.SetRequestMetadataTags(span, md)
-			instrumentation.SetTags(span, op.Tags())
+			grpctrace.SetRequestMetadataTags(span, md)
+			trace.SetTags(span, op.Tags())
 			if len(events) > 0 {
-				grpcsec.SetSecurityEventsTags(span, events)
+				grpctrace.SetSecurityEventsTags(span, events)
 			}
 		}()
 
@@ -64,16 +65,16 @@ func appsecUnaryHandlerMiddleware(span ddtrace.Span, handler grpc.UnaryHandler) 
 
 // StreamHandler wrapper to use when AppSec is enabled to monitor its execution.
 func appsecStreamHandlerMiddleware(span ddtrace.Span, handler grpc.StreamHandler) grpc.StreamHandler {
-	instrumentation.SetAppSecEnabledTags(span)
+	trace.SetAppSecEnabledTags(span)
 	return func(srv interface{}, stream grpc.ServerStream) error {
 		var err error
 		var blocked bool
 		ctx := stream.Context()
 		md, _ := metadata.FromIncomingContext(ctx)
 		clientIP := setClientIP(ctx, span, md)
-		grpcsec.SetRequestMetadataTags(span, md)
+		grpctrace.SetRequestMetadataTags(span, md)
 
-		ctx, op := grpcsec.StartHandlerOperation(ctx, grpcsec.HandlerOperationArgs{Metadata: md, ClientIP: clientIP}, nil, dyngo.NewDataListener(func(a *emitter.Action) {
+		ctx, op := grpcsec.StartHandlerOperation(ctx, grpcsec.HandlerOperationArgs{Metadata: md, ClientIP: clientIP}, nil, dyngo.NewDataListener(func(a *sharedsec.Action) {
 			code, e := a.GRPC()(md)
 			blocked = a.Blocking()
 			err = status.Error(codes.Code(code), e.Error())
@@ -86,11 +87,11 @@ func appsecStreamHandlerMiddleware(span ddtrace.Span, handler grpc.StreamHandler
 		defer func() {
 			events := op.Finish(grpcsec.HandlerOperationRes{})
 			if blocked {
-				op.AddTag(instrumentation.BlockedRequestTag, true)
+				op.AddTag(trace.BlockedRequestTag, true)
 			}
-			instrumentation.SetTags(span, op.Tags())
+			trace.SetTags(span, op.Tags())
 			if len(events) > 0 {
-				grpcsec.SetSecurityEventsTags(span, events)
+				grpctrace.SetSecurityEventsTags(span, events)
 			}
 		}()
 
@@ -131,10 +132,10 @@ func setClientIP(ctx context.Context, span ddtrace.Span, md metadata.MD) netip.A
 	if p, ok := peer.FromContext(ctx); ok {
 		remoteAddr = p.Addr.String()
 	}
-	ipTags, clientIP := httpsec.ClientIPTags(md, false, remoteAddr)
+	ipTags, clientIP := httptrace.ClientIPTags(md, false, remoteAddr)
 	log.Debug("appsec: http client ip detection returned `%s` given the http headers `%v`", clientIP, md)
 	if len(ipTags) > 0 {
-		instrumentation.SetStringTags(span, ipTags)
+		trace.SetTags(span, ipTags)
 	}
 	return clientIP
 }
