+++ internal/appsec/emitter/doc.go
@@ -1,0 +1,10 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016 Datadog, Inc.
+
+// Package emitter provides functions and types used to instrument go libraries
+// to integrate them with Datadog AppSec features. The events produced by this
+// instrumentation can be listened to using the functions found in
+// gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/listener.
+package emitter
