+++ internal/appsec/emitter/httpsec/tags_test.go
@@ -3,13 +3,14 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016 Datadog, Inc.
 
-package emitter
+package httpsec
 
 import (
 	"fmt"
 	"testing"
 
 	testlib "gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/_testlib"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/trace"
 
 	"github.com/stretchr/testify/require"
 )
@@ -101,7 +102,7 @@ func TestTags(t *testing.T) {
 				respHeadersCase := respHeadersCase
 				t.Run(fmt.Sprintf("%s-%s-%s", eventCase.name, reqHeadersCase.name, respHeadersCase.name), func(t *testing.T) {
 					var span testlib.MockSpan
-					err := setSecurityEventsTags(&span, eventCase.events)
+					err := trace.SetEventSpanTags(&span, eventCase.events)
 					if eventCase.expectedError {
 						require.Error(t, err)
 						return
@@ -131,44 +132,5 @@ func TestTags(t *testing.T) {
 				})
 			}
 		}
-	}
-}
-
-func TestNormalizeHTTPHeaders(t *testing.T) {
-	for _, tc := range []struct {
-		headers  map[string][]string
-		expected map[string]string
-	}{
-		{
-			headers:  nil,
-			expected: nil,
-		},
-		{
-			headers: map[string][]string{
-				"cookie": {"not-collected"},
-			},
-			expected: nil,
-		},
-		{
-			headers: map[string][]string{
-				"cookie":          {"not-collected"},
-				"x-forwarded-for": {"1.2.3.4,5.6.7.8"},
-			},
-			expected: map[string]string{
-				"x-forwarded-for": "1.2.3.4,5.6.7.8",
-			},
-		},
-		{
-			headers: map[string][]string{
-				"cookie":          {"not-collected"},
-				"x-forwarded-for": {"1.2.3.4,5.6.7.8", "9.10.11.12,13.14.15.16"},
-			},
-			expected: map[string]string{
-				"x-forwarded-for": "1.2.3.4,5.6.7.8,9.10.11.12,13.14.15.16",
-			},
-		},
-	} {
-		headers := NormalizeHTTPHeaders(tc.headers)
-		require.Equal(t, tc.expected, headers)
 	}
 }
