+++ internal/appsec/emitter/sharedsec/actions.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2022 Datadog, Inc.
 
-package emitter
+package sharedsec
 
 import (
 	_ "embed" // Blank import
