+++ internal/appsec/emitter/sharedsec/actions_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016 Datadog, Inc.
 
-package emitter
+package sharedsec
 
 import (
 	"io"
