+++ internal/appsec/emitter/sharedsec/blocked-template.html
