+++ internal/appsec/emitter/sharedsec/blocked-template.json
