+++ internal/appsec/emitter/sharedsec/shared.go
@@ -3,14 +3,14 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2023 Datadog, Inc.
 
-package emitter
+package sharedsec
 
 import (
 	"context"
 	"reflect"
 
 	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/dyngo"
-	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/dyngo/instrumentation"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/listener"
 	"gopkg.in/DataDog/dd-trace-go.v1/internal/log"
 )
 
@@ -61,7 +61,7 @@ func (f OnUserIDOperationStart) Call(op dyngo.Operation, v interface{}) {
 // A call to the WAF is made to check the user ID and an error is returned if the
 // user should be blocked. The return value is nil otherwise.
 func MonitorUser(ctx context.Context, userID string) error {
-	if parent, ok := ctx.Value(instrumentation.ContextKey{}).(dyngo.Operation); ok {
+	if parent, ok := ctx.Value(listener.ContextKey{}).(dyngo.Operation); ok {
 		return ExecuteUserIDOperation(parent, UserIDOperationArgs{UserID: userID})
 	}
 	log.Error("appsec: user ID monitoring ignored: could not find the http handler instrumentation metadata in the request context: the request handler is not being monitored by a middleware function or the provided context is not the expected request context")
